finally!!!!!!!!!!!
I made use of createElement() and append() methods. So HTML2JS was very helpful for converting html to js that way. so the most important thing was that I moved those element that I will add event listener to from a local scope to a global scope. I also needed to manually add some classes that those element added by the main event created but are not already owned by classes originally in the dom cos tailwind only add classes only in the dom. I don't know if insertAdjacentHTML or inner HTML would have worked with the whole declaring as a global var stuffs