just cleaning up the markup and adding some update .
still stuck on element not listening to event even when present in the dom issue. my method so far for solving it has been getElementsByClass() method and element[i]. this will certainly work for those element that I createElement() with but had to add another entire event listener for just by using classes and getElementsByClass() method.