started writing lines to make use of a boiler plate html file for all chats on <md screens instead of multiple html for each.
the next step will be to carefully integrate the use of the same div.chats that I'm currently using for md> screens so as to further clean up this code. I must also carefully rename all the instances of barber.html throughout the project. All in all, I am getting better at this web development journey.