fixed dark-mode-bg-not-filling the page on-mobile-due-to-height-of-slide-container-not-filling col-1 and some other bugs
fixed the input-anime-container not working on chat.html i.e. mobile, fixed the lagging settings dropdown on the chat-page on the right-column on md> screens, fixed both right and left columns (md>) settings-dropdown-not-hidding when clicking on #slide-container, and also fixed dropdown not hidding on scroll on mobile chatpage i.e. chat.html